+++ ui/desktop/src/config.ts
@@ -9,3 +9,33 @@ export const getApiUrl = (endpoint: string): string => {
 export const getSecretKey = (): string => {
   return window.appConfig.get('secretKey');
 }
+
+
+// Function to send the system configuration to the server
+export const addSystemConfig = async (system: string) => {
+  console.log("calling add system")
+  
+  const systemConfig = {
+    type: "Stdio",
+    cmd: await window.electron.getBinaryPath('goosed'),
+    args: ["mcp", system]
+  };
+
+  try {
+    const response = await fetch(getApiUrl('/systems/add'), {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(systemConfig)
+    });
+
+    if (!response.ok) {
+      throw new Error(`Failed to add system config for ${system}: ${response.statusText}`);
+    }
+
+    console.log(`Successfully added system config for ${system}`);
+  } catch (error) {
+    console.log(`Error adding system config for ${system}:`, error);
+  }
+};
